+++ src/components/atoms/Introduction.vue
@@ -1,344 +1,180 @@
 <template>
-  <section>
-    <!-- Desktop -->
-    <v-container
-      style="display:contents; background-color:#f4b700; height:300px"
-      class="hidden-sm-and-down"
-    >
-      <v-img src="@/assets/landing/introductionDesktop.svg">
-        <v-container>
-          <v-row align="center" class="mb-10">
-            <v-col class="text-left" cols="12" md="6">
-              <h1 class="display-3 font-weight-regular mb-4 white--text">
-                UX Remote LAB
-              </h1>
-              <h4 class="display-1 white--text mb-4">
-                What about doing Usability Tests in remote?
-              </h4>
-              <p class="white--text mb-4" style="width: 80%" align="justify">
-                Join our OpenSource project and start creating your own
-                usability LAB and share it with friends. Researching has never
-                been easier!
-              </p>
-              <v-btn
-                color="white"
-                outlined
-                rounded
-                class="mb-2"
-                @click="goTo('/signup')"
-                >Get started</v-btn
-              >
-            </v-col>
-          </v-row>
+    <section>
+        <!-- Desktop -->
+        <v-container
+            style="display:contents; background-color:#f4b700; height:300px"
+            class="hidden-sm-and-down"
+        >
+            <v-img src="@/assets/landing/introductionDesktop.svg">
+                <v-container>
+                    <v-row align="center" class="mb-10">
+                        <v-col class="text-left" cols="12" md="6">
+                            <h1
+                                class="display-3 font-weight-regular mb-4 white--text"
+                            >
+                                UX Remote LAB
+                            </h1>
+                            <h4 class="display-1 white--text mb-4">
+                                What about doing Usability Tests in remote?
+                            </h4>
+                            <p
+                                class="white--text mb-4"
+                                style="width: 80%"
+                                align="justify"
+                            >
+                                Join our OpenSource project and start creating
+                                your own usability LAB and share it with
+                                friends. Researching has never been easier!
+                            </p>
+                            <v-btn
+                                color="white"
+                                outlined
+                                rounded
+                                class="mb-2"
+                                @click="goTo('/signup')"
+                                >Get started</v-btn
+                            >
+                        </v-col>
+                    </v-row>
+                </v-container>
+            </v-img>
         </v-container>
-      </v-img>
-    </v-container>
 
-    <!-- Mobile -->
-    <v-container class="hidden-md-and-up ma-0 pa-0">
-      <div style="background-color: #f4b700">
-        <div style="background-color: #f4b700">
-          <h1 class="display-3 font-weight-regular white--text text-center">
-            UX Remote LAB
-          </h1>
-        </div>
-        <v-img
-          src="@/assets/landing/introductionMobile.svg"
-          class="mb-4"
-          max-height="350"
-          contain
-        ></v-img>
-        <div style="background-color: #f4b700" class="mx-1">
-          <h4 class="display-1 white--text mb-4 text-center">
-            What about doing Usability Tests in remote?
-          </h4>
-        </div>
-        <div style="background-color: #f4b700" class="mx-3">
-          <v-row justify="center">
-            <p class="white--text mb-4 mx-4 text-center">
-              Join our OpenSource project and start creating your own usability
-              LAB and share it with friends. Researching has never been easier!
-            </p>
-            <v-btn
-              color="white"
-              outlined
-              rounded
-              class="mb-2"
-              @click="goTo('/signup')"
-              >Get started</v-btn
+        <!-- Mobile -->
+        <v-container class="hidden-md-and-up ma-0 pa-0">
+            <div style="background-color: #f4b700">
+                <div style="background-color: #f4b700">
+                    <h1
+                        class="display-3 font-weight-regular white--text text-center"
+                    >
+                        UX Remote LAB
+                    </h1>
+                </div>
+                <v-img
+                    src="@/assets/landing/introductionMobile.svg"
+                    class="mb-4"
+                    max-height="350"
+                    contain
+                ></v-img>
+                <div style="background-color: #f4b700" class="mx-1">
+                    <h4 class="display-1 white--text mb-4 text-center">
+                        What about doing Usability Tests in remote?
+                    </h4>
+                </div>
+                <div style="background-color: #f4b700" class="mx-3">
+                    <v-row justify="center">
+                        <p class="white--text mb-4 mx-4 text-center">
+                            Join our OpenSource project and start creating your
+                            own usability LAB and share it with friends.
+                            Researching has never been easier!
+                        </p>
+                        <v-btn
+                            color="white"
+                            outlined
+                            rounded
+                            class="mb-2"
+                            @click="goTo('/signup')"
+                            >Get started</v-btn
+                        >
+                    </v-row>
+                </div>
+                <!-- div for margin at bottom -->
+                <div style="background-color: #f4b700; height: 40px"></div>
+            </div>
+
+            <div style="height: 10px"></div>
+        </v-container>
+
+        <div class="svg-border-waves text-white">
+            <svg
+                class="wave"
+                style="pointer-events: none"
+                fill="white"
+                preserveAspectRatio="none"
+                xmlns="http://www.w3.org/2000/svg"
+                xlink="http://www.w3.org/1999/xlink"
+                viewBox="0 0 1920 75"
             >
-          </v-row>
+                <defs>
+                    <clipPath id="a">
+                        <rect class="a" width="1920" height="75" />
+                    </clipPath>
+                </defs>
+                <title>wave</title>
+                <g class="b">
+                    <path
+                        class="c"
+                        d="M1963,327H-105V65A2647.49,2647.49,0,0,1,431,19c217.7,3.5,239.6,30.8,470,36,297.3,6.7,367.5-36.2,642-28a2511.41,2511.41,0,0,1,420,48"
+                    />
+                </g>
+                <g class="b">
+                    <path
+                        class="d"
+                        d="M-127,404H1963V44c-140.1-28-343.3-46.7-566,22-75.5,23.3-118.5,45.9-162,64-48.6,20.2-404.7,128-784,0C355.2,97.7,341.6,78.3,235,50,86.6,10.6-41.8,6.9-127,10"
+                    />
+                </g>
+                <g class="b">
+                    <path
+                        class="d"
+                        d="M1979,462-155,446V106C251.8,20.2,576.6,15.9,805,30c167.4,10.3,322.3,32.9,680,56,207,13.4,378,20.3,494,24"
+                    />
+                </g>
+                <g class="b hidden-sm-and-down">
+                    <path
+                        class="d"
+                        d="M1998,484H-243V100c445.8,26.8,794.2-4.1,1035-39,141-20.4,231.1-40.1,378-45,349.6-11.6,636.7,73.8,828,150"
+                    />
+                </g>
+            </svg>
         </div>
-        <!-- div for margin at bottom -->
-        <div style="background-color: #f4b700; height: 40px"></div>
-      </div>
-
-      <div style="height: 10px"></div>
-    </v-container>
-
-    <div class="svg-border-waves text-white">
-      <svg
-        class="wave"
-        style="pointer-events: none"
-        fill="white"
-        preserveAspectRatio="none"
-        xmlns="http://www.w3.org/2000/svg"
-        xlink="http://www.w3.org/1999/xlink"
-        viewBox="0 0 1920 75"
-      >
-        <defs>
-          <clipPath id="a">
-            <rect class="a" width="1920" height="75" />
-          </clipPath>
-        </defs>
-        <title>wave</title>
-        <g class="b">
-          <path
-            class="c"
-            d="M1963,327H-105V65A2647.49,2647.49,0,0,1,431,19c217.7,3.5,239.6,30.8,470,36,297.3,6.7,367.5-36.2,642-28a2511.41,2511.41,0,0,1,420,48"
-          />
-        </g>
-        <g class="b">
-          <path
-            class="d"
-            d="M-127,404H1963V44c-140.1-28-343.3-46.7-566,22-75.5,23.3-118.5,45.9-162,64-48.6,20.2-404.7,128-784,0C355.2,97.7,341.6,78.3,235,50,86.6,10.6-41.8,6.9-127,10"
-          />
-        </g>
-        <g class="b">
-          <path
-            class="d"
-            d="M1979,462-155,446V106C251.8,20.2,576.6,15.9,805,30c167.4,10.3,322.3,32.9,680,56,207,13.4,378,20.3,494,24"
-          />
-        </g>
-        <g class="b hidden-sm-and-down">
-          <path
-            class="d"
-            d="M1998,484H-243V100c445.8,26.8,794.2-4.1,1035-39,141-20.4,231.1-40.1,378-45,349.6-11.6,636.7,73.8,828,150"
-          />
-        </g>
-      </svg>
-    </div>
-
-    <div id="FileUpload">
-      <v-row justify="center">
-        <v-col class="ma-10" cols="10">
-          <v-row class="ma-2" justify="center" align="center">
-            <v-file-input
-              v-model="csvFile"
-              class="d-flex justify-center "
-              accept=".csv"
-              show-size
-              truncate-length="15"
-              placeholder="Import your CSV file here."
-              ref="myFile"
-            >
-            </v-file-input>
-            <v-btn
-              :loading="loading"
-              :disabled="loading"
-              color="blue-grey"
-              class="ma-2 white--text"
-              @click="(loader = 'loading'), changeToJSON()"
-            >
-              Upload
-              <v-icon right dark>
-                mdi-cloud-upload
-              </v-icon>
-            </v-btn>
-          </v-row>
-        </v-col>
-      </v-row>
-    </div>
-  </section>
+    </section>
 </template>
 
 <style scoped>
 .a {
-  fill: none;
+    fill: none;
 }
 .b {
-  clip-path: url(#a);
+    clip-path: url(#a);
 }
 .d {
-  opacity: 0.5;
-  isolation: isolate;
+    opacity: 0.5;
+    isolation: isolate;
 }
 .svg-border-waves svg {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  height: 3rem;
-  width: 100%;
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    height: 3rem;
+    width: 100%;
 }
 svg {
-  overflow: hidden;
+    overflow: hidden;
 }
 section {
-  position: relative;
+    position: relative;
 }
 </style>
 
 <script>
 export default {
-  data: () => ({
-    //remover
-    loading: false,
-    loader: null,
-    csvFile: null,
-    heuristicForm: null,
-    refs: this.$refs,
-  }),
-  methods: {
-    goTo(path) {
-      this.$router.push(path).catch(() => {});
+    data: () => ({}),
+    methods: {
+        goTo(path) {
+            this.$router.push(path).catch(() => {});
+        },
     },
-    //remover
-    async changeToJSON() {
-      const testId = "JArhiG9o6yFTy76sDtve";
-
-      console.log(testId);
-
-      let lines = "";
-      let currentline = "";
-      let csv = "";
-      let headers = "";
-      let result = [];
-      let result2 = [];
-      let reader = new FileReader();
-
-      reader.readAsBinaryString(this.csvFile);
-
-      reader.onload = async () => {
-        csv = reader.result;
-        lines = csv.split("\r" + "\n");
-        headers = lines[0].split(";");
-        headers[0] = headers[0].slice(3);
-
-        for (var i = 1; i < lines.length; i++) {
-          if (!lines[i]) continue;
-          let obj = {};
-          currentline = lines[i];
-          var re = /"/g;
-          currentline = re[Symbol.replace](currentline, "");
-          currentline = currentline.split(";");
-
-          for (var j = 0; j < headers.length; j++) {
-            if (j == 0 || j == 1 || j == 2 || j == 3) {
-              let head = headers[j];
-              let value = currentline[j];
-              obj[head] = value;
-            }
-          }
-          result.push(obj);
-        }
-        result = JSON.stringify(result);
-        result2 = JSON.parse(result.toString());
-
-        function getQuestionsFromHeuristic(heuristicArray, heuristicId) {
-          return heuristicArray.filter((element) => element.HID == heuristicId);
-        }
-        let heuristicTest = [];
-        let heuristicOcurrencies = 0;
-
-        for (i = 0; i < result2.length; i++) {
-          let auxHeuristic = getQuestionsFromHeuristic(result2, i + 1);
-          if (auxHeuristic.length > 0) {
-            heuristicOcurrencies = heuristicOcurrencies + 1;
-            let auxQuestions = [];
-            for (j = 0; j < auxHeuristic.length; j++) {
-              let auxQuestion = new HeuristicQuestion(
-                auxHeuristic[j].QID,
-                auxHeuristic[j].QUESTION,
-                auxHeuristic[j].QUESTION,
-                auxHeuristic[j].QUESTION
-              );
-              auxQuestions.push(auxQuestion);
-            }
-            let setHeuristics = new Heuristic(
-              heuristicOcurrencies,
-              auxHeuristic[0].HEURISTIC,
-              auxQuestions,
-              auxQuestions.length
-            );
-            heuristicTest.push(setHeuristics);
-            console.log(setHeuristics);
-          }
-        }
-
-        console.log(this.test);
-
-        this.$store.dispatch("saveCurrentTest", heuristicTest);
-        console.log(testId);
-        console.log(heuristicTest);
-
-        // await new HeuristicController().createCsvHeuris({
-        //   // testId: testId,
-        //   // id: heuristicTest[0].id,
-        //   // qd: heuristicTest[0].questions[0].descriptions,
-        //   // qid: heuristicTest[0].questions[0].id,
-        //   // qtext: heuristicTest[0].questions[0].text,
-        //   // qtitle: heuristicTest[0].questions[0].title,
-        //   // title: heuristicTest[0].title,
-        //   // total: heuristicTest[0].total,
-        // });
-        console.log(heuristicTest);
-
-        for (i = 0; i < heuristicTest.length; i++) {
-          console.log(heuristicTest[i].id);
-          let aux = Array.of(heuristicTest[i].total);
-          for (j = 0; j < heuristicTest[i].total; j++) {
-            console.log(heuristicTest[i].questions[j]);
-
-            aux[j] = {
-              id: heuristicTest[i].questions[j].id,
-              res: heuristicTest[i].questions[j].title,
-              com: heuristicTest[i].questions[j].descriptions,
-            };
-            console.log("auxiliar: " + aux);
-          }
-          await new HeuristicController().createCsvHeuris({
-            testId: testId,
-            id: heuristicTest[i].id,
-            questions: aux,
-            title: heuristicTest[i].title,
-            total: heuristicTest[i].total,
-          });
-        }
-      };
+    computed: {
+        test() {
+            return this.$store.getters.test;
+        },
+        user() {
+            return this.$store.getters.user;
+        },
+        csvHeuristics() {
+            return this.$store.state.Tests.currentTest;
+        },
     },
-  },
-  watch: {
-    loader() {
-      const l = this.loader;
-      this[l] = !this[l];
-      // const alertFunc = alert("Your file has been uploaded!");
-
-      if (this.csvFile != null) {
-        setTimeout(() => (this[l] = false), 3000);
-        setTimeout(() => (this.csvFile = null), 3000);
-        // setTimeout(alertFunc, 3000);
-        this.loader = null;
-      } else {
-        setTimeout(() => (this[l] = false), 3000);
-        alert("No csv file selected. \nPlease select one before procede.");
-        this.loader = null;
-      }
-    },
-  },
-  computed: {
-    test() {
-      return this.$store.getters.test;
-    },
-    user() {
-      return this.$store.getters.user;
-    },
-    csvHeuristics() {
-      return this.$store.state.Tests.currentTest;
-    },
-  },
 };
 //remover
-import Heuristic from "@/models/Heuristic";
-import HeuristicQuestion from "@/models/HeuristicQuestion";
-import HeuristicController from "../../controllers/HeuristicController";
 </script>
